+++ src/main/kotlin/com/maedjyukghoti/tictactoe/display/UserInterface.kt
@@ -1,8 +1,7 @@
 package com.maedjyukghoti.tictactoe.display
 
 import com.github.michaelbull.result.*
-import com.maedjyukghoti.tictactoe.AppState
-import com.maedjyukghoti.tictactoe.UserIntent
+import com.maedjyukghoti.tictactoe.*
 import com.maedjyukghoti.tictactoe.logic.*
 import kotlinx.coroutines.*
 import kotlinx.coroutines.flow.*
