+++ src/main/kotlin/com/maedjyukghoti/tictactoe/logic/players/BotStrategy.kt
@@ -1,33 +1,31 @@
 package com.maedjyukghoti.tictactoe.logic.players
 
 import com.github.michaelbull.result.*
+import com.maedjyukghoti.tictactoe.MoveError
 import com.maedjyukghoti.tictactoe.logic.*
 
 sealed interface BotStrategy {
     fun getCoordinates(gameState: GameState, botPlayer: PlayerInfo): Result<Coordinates, MoveError>
 
     companion object {
-        fun getBotAtLevel(botLevel: Int): BotStrategy {
-            return when (botLevel) {
+        fun getBotAtLevel(botLevel: Int): BotStrategy =
+            when (botLevel) {
                 1 -> OneLayer
                 2 -> TwoLayer
                 3 -> MiniMax
                 else -> Random
             }
-        }
     }
 
-    object Random : BotStrategy {
-        override fun getCoordinates(gameState: GameState, botPlayer: PlayerInfo): Result<Coordinates, MoveError> {
-            return getRandomCoordinates(gameState.board)
-        }
+    data object Random : BotStrategy {
+        override fun getCoordinates(gameState: GameState, botPlayer: PlayerInfo): Result<Coordinates, MoveError> =
+            getRandomCoordinates(gameState.board)
     }
 
-    object OneLayer : BotStrategy {
-        override fun getCoordinates(gameState: GameState, botPlayer: PlayerInfo): Result<Coordinates, MoveError> {
-            return getWinningCoordinates(gameState.board, botPlayer)
+    data object OneLayer : BotStrategy {
+        override fun getCoordinates(gameState: GameState, botPlayer: PlayerInfo): Result<Coordinates, MoveError> =
+            getWinningCoordinates(gameState.board, botPlayer)
                 .orElse { getRandomCoordinates(gameState.board) }
-        }
     }
 
     /**
@@ -37,52 +35,53 @@ sealed interface BotStrategy {
      *  2) first to have a blocking move returns
      *  3) else no blocking move
      */
-    object TwoLayer : BotStrategy {
-        override fun getCoordinates(gameState: GameState, botPlayer: PlayerInfo): Result<Coordinates, MoveError> {
-            return getWinningCoordinates(gameState.board, botPlayer)
+    data object TwoLayer : BotStrategy {
+        override fun getCoordinates(gameState: GameState, botPlayer: PlayerInfo): Result<Coordinates, MoveError> =
+            getWinningCoordinates(gameState.board, botPlayer)
                 .orElse { getBlockingCoordinates(gameState.board, PlayerInfo.nextPlayer(botPlayer)) }
                 .orElse { getRandomCoordinates(gameState.board) }
-        }
     }
 
-    object MiniMax : BotStrategy {
+    data object MiniMax : BotStrategy {
         private val memo: MutableMap<Set<MoveRequest>, Choice> = mutableMapOf()
-        override fun getCoordinates(gameState: GameState, botPlayer: PlayerInfo): Result<Coordinates, MoveError> {
-            val choice = minimax(gameState.board, botPlayer, gameState.currentPlayerInfo, 0, memo)
-            return Ok(choice.option)
-        }
+        override fun getCoordinates(gameState: GameState, botPlayer: PlayerInfo): Result<Coordinates, MoveError> =
+            Ok(
+                minimax(
+                    board = gameState.board,
+                    myPlayer = botPlayer,
+                    currentPlayer = gameState.currentPlayerInfo,
+                    depth = 0,
+                    memo = memo
+                ).option
+            )
     }
 }
 
-fun getRandomCoordinates(board: Board): Result<Coordinates, MoveError> {
-    return runCatching { board.getRemainingCoordinates().random() }
+fun getRandomCoordinates(board: Board): Result<Coordinates, MoveError> =
+    runCatching { board.getRemainingCoordinates().random() }
         .mapError { MoveError.NoAvailableMoves }
-}
 
 /** Return the [Coordinates] to block [PlayerInfo] if they exist **/
-fun getBlockingCoordinates(board: Board, playerInfo: PlayerInfo): Result<Coordinates, MoveError> {
-    return getWinningCoordinates(board, playerInfo)
-}
+fun getBlockingCoordinates(board: Board, playerInfo: PlayerInfo): Result<Coordinates, MoveError> =
+    getWinningCoordinates(board, playerInfo)
 
 /** Return the winning [Coordinates] for [PlayerInfo] on the [Board] if they exist. **/
-fun getWinningCoordinates(board: Board, playerInfo: PlayerInfo): Result<Coordinates, MoveError> {
-    return runCatching {
+fun getWinningCoordinates(board: Board, playerInfo: PlayerInfo): Result<Coordinates, MoveError> =
+    runCatching {
         board.getRemainingCoordinates()
             .first { coordinates ->
                 val moveRequest = MoveRequest(coordinates, playerInfo)
-                val newBoard = makeMove(moveRequest, board)
+                val newBoard = board.makeMove(moveRequest)
                 val winner = checkForWinner(newBoard)
                 winner == playerInfo
             }
     }.mapError { MoveError.NoWinningMove }
-}
 
 data class Choice(val option: Coordinates, val value: Int, val depth: Int) : Comparable<Choice> {
-    override fun compareTo(other: Choice) : Int {
-        return if (value > other.value) 1
+    override fun compareTo(other: Choice) : Int =
+        if (value > other.value) 1
         else if (value < other.value) -1
         else 0
-    }
 }
 
 fun minimax(board: Board, myPlayer: PlayerInfo, currentPlayer: PlayerInfo, depth: Int, memo: MutableMap<Set<MoveRequest>, Choice>): Choice {
@@ -97,7 +96,7 @@ fun minimax(board: Board, myPlayer: PlayerInfo, currentPlayer: PlayerInfo, depth
         board.getRemainingCoordinates()
             .map { coordinates ->
                 val moveRequest = MoveRequest(coordinates, currentPlayer)
-                val updateBoard = makeMove(moveRequest, board)
+                val updateBoard = board.makeMove(moveRequest)
                 val nextPlayer = PlayerInfo.nextPlayer(currentPlayer)
                 val result = minimax(updateBoard, myPlayer, nextPlayer, depth + 1, memo)
                 result.copy(option = updateBoard.moves.last().coordinates)
