+++ src/main/kotlin/com/maedjyukghoti/tictactoe/logic/players/Player.kt
@@ -1,6 +1,7 @@
 package com.maedjyukghoti.tictactoe.logic.players
 
 import com.github.michaelbull.result.*
+import com.maedjyukghoti.tictactoe.UserIntent
 import com.maedjyukghoti.tictactoe.logic.*
 
 sealed interface Player {
@@ -10,9 +11,12 @@ sealed interface Player {
         private val playerInfo: PlayerInfo,
         private val botStrategy: BotStrategy
     ) : Player {
-        fun getAction(gameState: GameState): Result<Action, GameError> {
+        fun getAction(gameState: GameState): UserIntent {
             return botStrategy.getCoordinates(gameState, playerInfo)
-                .map { coordinates -> Action.Move(coordinates) }
+                .fold(
+                    success = { coordinates -> UserIntent.Move(coordinates) },
+                    failure = { moveError -> UserIntent.Error(moveError) }
+                )
         }
     }
 }
