+++ src/test/kotlin/MakeMoveTest.kt
@@ -1,7 +1,6 @@
 import com.maedjyukghoti.tictactoe.logic.Coordinates
 import com.maedjyukghoti.tictactoe.logic.MoveRequest
 import com.maedjyukghoti.tictactoe.logic.PlayerInfo
-import com.maedjyukghoti.tictactoe.logic.makeMove
 import org.junit.Test
 import kotlin.test.assertTrue
 
@@ -10,7 +9,7 @@ internal class MakeMoveTest {
     fun `move on board increases move count by 1`() {
         val oldBoard = getEmptyBoard(3)
         val request = MoveRequest(Coordinates(1, 1), PlayerInfo.Two)
-        val newBoard = makeMove(request, oldBoard)
+        val newBoard = oldBoard.makeMove(request)
 
         assertTrue(oldBoard.moves.size < newBoard.moves.size, "newBoard should be larger than old board")
         assertTrue(newBoard.moves.containsAll(oldBoard.moves), "oldBoard should be a subset of newBoard")
